Now works properly on non-localhosts.
I made a farily bad assumption when designing the code originally that
the session variable would be available and the same throughout the
entire process. That is obviously not true since the requests sometimes
come from different servers.

Anyway, it now uses Memcache (hopefully) or a FakeCache (if you're not
in a "production" RACK_ENV.

Also added some error messages if people don't use correct credentails.